Refactored tests to make things fit better with the new PyJWT object API - test_jwt.py has been renamed to test_api.py since it focuses entirely   on api.py - A new test_jwt.py has been introduced that focuses exclusively on   testing the public API. (Specifically, making sure that encode and   decode still exist and function). This test can be extremely simple   since the jwt.encode and jwt.decode functions are backed by PyJWT()   which is tested elsewhere.
